+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [vNext]
 - [C,D] `pallet-ddc-customers`: implemented bucket removal
 - Added ChargeError event to payout pallet
+- [C,D] Remove pallet-society
 
 - ...